Extracted Interfaces and broke out multiple items into their own files and also made a base record for most UploadRequest items
