+++ src/CatBox.NET/CatBoxClient.cs
@@ -1,95 +1,8 @@
-using System.Runtime.CompilerServices;
+﻿using System.Runtime.CompilerServices;
 using Microsoft.Extensions.Options;
 using static CatBox.NET.Common;
 
 namespace CatBox.NET;
-
-public interface ICatBoxClient
-{
-    /// <summary>
-    /// Enables uploading multiple files from disk (FileStream) to the API
-    /// </summary>
-    /// <param name="fileUploadRequest"></param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException">When <see cref="FileUploadRequest"/> is null</exception>
-    /// <returns>Response string from the API</returns>
-    IAsyncEnumerable<string?> UploadMultipleImages(FileUploadRequest fileUploadRequest, CancellationToken ct = default);
-
-    /// <summary>
-    /// Enables uploading multiple files by URL to the API
-    /// </summary>
-    /// <param name="urlUploadRequest">Data to send to the API</param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException">When <see cref="UrlUploadRequest"/> is null</exception>
-    /// <exception cref="HttpRequestException"> when something bad happens when talking to the API</exception>
-    /// <returns>Response string from the API</returns>
-    IAsyncEnumerable<string?> UploadMultipleUrls(UrlUploadRequest urlUploadRequest, CancellationToken ct = default);
-
-    /// <summary>
-    /// Deletes multiple files by API file name
-    /// </summary>
-    /// <param name="deleteFileRequest">Files to delete from the server</param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException">When <see cref="DeleteFileRequest"/> is null</exception>
-    /// <exception cref="ArgumentNullException">When <see cref="DeleteFileRequest.UserHash"/> is null</exception>
-    /// <exception cref="ArgumentException">When <see cref="DeleteFileRequest.FileNames"/> is null, empty, or whitespace</exception>
-    /// <exception cref="HttpRequestException"> when something bad happens when talking to the API</exception>
-    /// <returns>Response string from the API</returns>
-    Task<string?> DeleteMultipleFiles(DeleteFileRequest deleteFileRequest, CancellationToken ct = default);
-
-    /// <summary>
-    /// Streams a single image to be uploaded
-    /// </summary>
-    /// <param name="fileUploadRequest"></param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException">When <see cref="StreamUploadRequest"/> is null</exception>
-    /// <exception cref="ArgumentNullException">When <see cref="StreamUploadRequest.FileName"/> is null</exception>
-    /// <exception cref="HttpRequestException"> when something bad happens when talking to the API</exception>
-    /// <returns>Response string from the API</returns>
-    Task<string?> UploadImage(StreamUploadRequest fileUploadRequest, CancellationToken ct = default);
-
-    /// <summary>
-    /// Creates an album on CatBox via provided file names generated by the API
-    /// </summary>
-    /// <param name="createAlbumRequest">Data to pass to the API</param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException"> when <see cref="CreateAlbumRequest"/> is null</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="CreateAlbumRequest.Description"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="CreateAlbumRequest.Title"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="CreateAlbumRequest.Files"/> is null, empty, or whitespace</exception>
-    /// <exception cref="HttpRequestException"> when something bad happens when talking to the API</exception>
-    /// <returns>Response string from the API</returns>
-    Task<string?> CreateAlbum(CreateAlbumRequest createAlbumRequest, CancellationToken ct = default);
-
-    /// <summary>
-    /// Edits the content of album according to the content that is passed to the API
-    /// </summary>
-    /// <param name="editAlbumRequest">Data to pass to the API</param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException"> when <see cref="EditAlbumRequest"/> is null</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="EditAlbumRequest.UserHash"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="EditAlbumRequest.Description"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="EditAlbumRequest.Title"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="EditAlbumRequest.AlbumId"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="EditAlbumRequest.Files"/> is null, empty, or whitespace</exception>
-    /// <exception cref="HttpRequestException"> when something bad happens when talking to the API</exception>
-    /// <returns>Response string from the API</returns>
-    Task<string?> EditAlbum(EditAlbumRequest editAlbumRequest, CancellationToken ct = default);
-
-    /// <summary>
-    /// This endpoint is for adding files to an album, removing files from an album, or deleting the album
-    /// </summary>
-    /// <param name="albumRequest">Data to pass to the API</param>
-    /// <param name="ct">Cancellation Token</param>
-    /// <exception cref="ArgumentNullException"> when <see cref="AlbumRequest"/> </exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="AlbumRequest.UserHash"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentNullException"> when <see cref="AlbumRequest.Files"/> is null, empty, or whitespace</exception>
-    /// <exception cref="ArgumentException"> when <see cref="AlbumRequest.Request"/> is not valid for this request type</exception>
-    /// <exception cref="InvalidOperationException"> when <see cref="AlbumRequest.Request"/> is not CatBoxRequestTypes.AddToAlbum, CatBoxRequestTypes.RemoveFromAlbum, CatBoxRequestTypes.DeleteAlbum</exception>
-    /// <exception cref="HttpRequestException"> when something bad happens when talking to the API</exception>
-    /// <returns>Response string from the API</returns>
-    Task<string?> ModifyAlbum(AlbumRequest albumRequest, CancellationToken ct = default);
-}
 
 public class CatBoxClient : ICatBoxClient
 {
@@ -97,10 +10,10 @@ public class CatBoxClient : ICatBoxClient
     private readonly CatBoxConfig _config;
 
     /// <summary>
-    /// 
+    /// Creates a new <see cref="CatBoxClient"/>
     /// </summary>
-    /// <param name="client"></param>
-    /// <param name="config"></param>
+    /// <param name="client"><see cref="HttpClient"/></param>
+    /// <param name="config"><see cref="IOptions{TOptions}"/></param>
     /// <exception cref="ArgumentNullException"></exception>
     public CatBoxClient(HttpClient client, IOptions<CatBoxConfig> config)
     {
@@ -133,9 +46,7 @@ public class CatBoxClient : ICatBoxClient
             request.Content = content;
 
             using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-            // response.EnsureSuccessStatusCode();
-
-            yield return await response.Content.ReadAsStringAsync();
+            yield return await response.Content.ReadAsStringAsync(ct);
         }
     }
     
@@ -149,9 +60,10 @@ public class CatBoxClient : ICatBoxClient
             throw new ArgumentNullException(nameof(fileUploadRequest.FileName), "Argument cannot be null");
 
         using var request = new HttpRequestMessage(HttpMethod.Post, _config.CatBoxUrl);
-        using var content = new MultipartFormDataContent();
-
-        content.Add(new StringContent(CatBoxRequestTypes.UploadFile.ToRequest()), CatBoxRequestStrings.RequestType);
+        using var content = new MultipartFormDataContent
+        {
+            { new StringContent(CatBoxRequestTypes.UploadFile.ToRequest()), CatBoxRequestStrings.RequestType }
+        };
 
         if (!string.IsNullOrWhiteSpace(fileUploadRequest.UserHash))
             content.Add(new StringContent(fileUploadRequest.UserHash), CatBoxRequestStrings.UserHashType);
@@ -160,9 +72,7 @@ public class CatBoxClient : ICatBoxClient
         request.Content = content;
 
         using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-        // response.EnsureSuccessStatusCode();
-
-        return await response.Content.ReadAsStringAsync();
+        return await response.Content.ReadAsStringAsync(ct);
     }
 
     /// <inheritdoc/>
@@ -188,9 +98,7 @@ public class CatBoxClient : ICatBoxClient
             request.Content = content;
 
             using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-            // response.EnsureSuccessStatusCode();
-
-            yield return await response.Content.ReadAsStringAsync();
+            yield return await response.Content.ReadAsStringAsync(ct);
         }
     }
 
@@ -208,17 +116,16 @@ public class CatBoxClient : ICatBoxClient
             throw new ArgumentNullException(nameof(deleteFileRequest.FileNames), "File list cannot be empty");
 
         using var request = new HttpRequestMessage(HttpMethod.Post, _config.CatBoxUrl);
-        using var content = new MultipartFormDataContent();
-
-        content.Add(new StringContent(CatBoxRequestTypes.DeleteFile.ToRequest()), CatBoxRequestStrings.RequestType);
-        content.Add(new StringContent(deleteFileRequest.UserHash), CatBoxRequestStrings.UserHashType);
-        content.Add(new StringContent(fileNames), CatBoxRequestStrings.FileType);
+        using var content = new MultipartFormDataContent
+        {
+            { new StringContent(CatBoxRequestTypes.DeleteFile.ToRequest()), CatBoxRequestStrings.RequestType },
+            { new StringContent(deleteFileRequest.UserHash), CatBoxRequestStrings.UserHashType },
+            { new StringContent(fileNames), CatBoxRequestStrings.FileType }
+        };
         request.Content = content;
 
         using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-        // response.EnsureSuccessStatusCode();
-
-        return await response.Content.ReadAsStringAsync();
+        return await response.Content.ReadAsStringAsync(ct);
     }
     
     /// <inheritdoc/>
@@ -237,7 +144,7 @@ public class CatBoxClient : ICatBoxClient
 
         var links = createAlbumRequest.Files.Select(link =>
         {
-            if (link.Contains(_config.CatBoxUrl.Host))
+            if (link.Contains(_config.CatBoxUrl!.Host))
             {
                 return new Uri(link).PathAndQuery[1..];
             }
@@ -250,8 +157,10 @@ public class CatBoxClient : ICatBoxClient
             throw new ArgumentNullException(nameof(createAlbumRequest.Files), "File list cannot be empty");
 
         using var request = new HttpRequestMessage(HttpMethod.Post, _config.CatBoxUrl);
-        using var content = new MultipartFormDataContent();
-        content.Add(new StringContent(CatBoxRequestTypes.CreateAlbum.ToRequest()), CatBoxRequestStrings.RequestType);
+        using var content = new MultipartFormDataContent
+        {
+            { new StringContent(CatBoxRequestTypes.CreateAlbum.ToRequest()), CatBoxRequestStrings.RequestType }
+        };
 
         if (!string.IsNullOrWhiteSpace(createAlbumRequest.UserHash))
             content.Add(new StringContent(createAlbumRequest.UserHash), CatBoxRequestStrings.UserHashType);
@@ -265,9 +174,7 @@ public class CatBoxClient : ICatBoxClient
         request.Content = content;
 
         using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-        // response.EnsureSuccessStatusCode();
-
-        return await response.Content.ReadAsStringAsync();
+        return await response.Content.ReadAsStringAsync(ct);
     }
     
     /// <inheritdoc/>
@@ -297,20 +204,19 @@ public class CatBoxClient : ICatBoxClient
             throw new ArgumentNullException(nameof(editAlbumRequest.Files), "File list cannot be empty");
 
         using var request = new HttpRequestMessage(HttpMethod.Post, _config.CatBoxUrl);
-        using var content = new MultipartFormDataContent();
-
-        content.Add(new StringContent(CatBoxRequestTypes.EditAlbum.ToRequest()), CatBoxRequestStrings.RequestType);
-        content.Add(new StringContent(editAlbumRequest.UserHash), CatBoxRequestStrings.UserHashType);
-        content.Add(new StringContent(editAlbumRequest.AlbumId), CatBoxRequestStrings.AlbumIdShortType);
-        content.Add(new StringContent(editAlbumRequest.Title), CatBoxRequestStrings.TitleType);
-        content.Add(new StringContent(editAlbumRequest.Description), CatBoxRequestStrings.DescriptionType);
-        content.Add(new StringContent(fileNames), CatBoxRequestStrings.FileType);
+        using var content = new MultipartFormDataContent
+        {
+            { new StringContent(CatBoxRequestTypes.EditAlbum.ToRequest()), CatBoxRequestStrings.RequestType },
+            { new StringContent(editAlbumRequest.UserHash), CatBoxRequestStrings.UserHashType },
+            { new StringContent(editAlbumRequest.AlbumId), CatBoxRequestStrings.AlbumIdShortType },
+            { new StringContent(editAlbumRequest.Title), CatBoxRequestStrings.TitleType },
+            { new StringContent(editAlbumRequest.Description), CatBoxRequestStrings.DescriptionType },
+            { new StringContent(fileNames), CatBoxRequestStrings.FileType }
+        };
         request.Content = content;
 
         using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-        // response.EnsureSuccessStatusCode();
-
-        return await response.Content.ReadAsStringAsync();
+        return await response.Content.ReadAsStringAsync(ct);
     }
     
     /// <inheritdoc/>
@@ -320,7 +226,9 @@ public class CatBoxClient : ICatBoxClient
             throw new ArgumentNullException(nameof(albumRequest), "Argument cannot be null");
 
         if (IsAlbumRequestTypeValid(albumRequest))
+#pragma warning disable CA2208 // Instantiate argument exceptions correctly
             throw new ArgumentException("Invalid Request Type for album endpoint", nameof(albumRequest.Request));
+#pragma warning restore CA2208 // Instantiate argument exceptions correctly
 
         if (string.IsNullOrWhiteSpace(albumRequest.UserHash))
             throw new ArgumentNullException(nameof(albumRequest.UserHash),
@@ -340,10 +248,12 @@ public class CatBoxClient : ICatBoxClient
             throw new ArgumentNullException(nameof(albumRequest.Files), "File list cannot be empty");
 
         using var request = new HttpRequestMessage(HttpMethod.Post, _config.CatBoxUrl);
-        using var content = new MultipartFormDataContent();
-        content.Add(new StringContent(albumRequest.Request.ToRequest()), CatBoxRequestStrings.RequestType);
-        content.Add(new StringContent(albumRequest.UserHash), CatBoxRequestStrings.UserHashType);
-        content.Add(new StringContent(albumRequest.AlbumId), CatBoxRequestStrings.AlbumIdShortType);
+        using var content = new MultipartFormDataContent
+        {
+            { new StringContent(albumRequest.Request.ToRequest()), CatBoxRequestStrings.RequestType },
+            { new StringContent(albumRequest.UserHash), CatBoxRequestStrings.UserHashType },
+            { new StringContent(albumRequest.AlbumId), CatBoxRequestStrings.AlbumIdShortType }
+        };
 
         // If request type is AddToAlbum or RemoveFromAlbum
         if (albumRequest.Request is CatBoxRequestTypes.AddToAlbum or CatBoxRequestTypes.RemoveFromAlbum)
@@ -352,9 +262,7 @@ public class CatBoxClient : ICatBoxClient
         request.Content = content;
 
         using var response = await _client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, ct);
-        // response.EnsureSuccessStatusCode();
-
-        return await response.Content.ReadAsStringAsync();
+        return await response.Content.ReadAsStringAsync(ct);
     }
 
     /// <summary>
