+++ src/CatBox.NET/ExpireAfter.cs
@@ -1,0 +1,28 @@
+﻿namespace CatBox.NET;
+
+/// <summary>
+/// Image expiry in litterbox.moe
+/// </summary>
+public enum ExpireAfter
+{
+    
+    /// <summary>
+    /// Expire after 1 hour
+    /// </summary>
+    OneHour,
+    
+    /// <summary>
+    /// Expire after 12 hours
+    /// </summary>
+    TwelveHours,
+    
+    /// <summary>
+    /// Expire after one day (24 hours)
+    /// </summary>
+    OneDay,
+    
+    /// <summary>
+    /// Expire after three days (72 hours)
+    /// </summary>
+    ThreeDays
+}
