+++ src/CatBox.NET/UploadHost.cs
@@ -1,0 +1,10 @@
+﻿namespace CatBox.NET;
+
+/// <summary>
+/// Currently unused
+/// </summary>
+public enum UploadHost
+{
+    CatBox,
+    LitterBox
+}
